@@ -1,61 +1,68 @@
 module github.com/rancher/rancher
 
-go 1.19
+go 1.20
 
 // on release remove this wrangler replace and use the latest tag
-replace github.com/rancher/wrangler v1.1.1 => github.com/rancher/wrangler v1.1.1-0.20230705223603-201b4da5bdaf
+replace github.com/rancher/wrangler v1.1.1 => github.com/rancher/wrangler v1.1.1-0.20230831050635-df1bd5aae9df
 
 replace (
-	github.com/docker/distribution => github.com/docker/distribution v2.8.2+incompatible // oras dep requires a replace is set
-	github.com/docker/docker => github.com/docker/docker v20.10.25+incompatible // oras dep requires a replace is set
+	github.com/containerd/containerd => github.com/containerd/containerd v1.6.22 // for compatibilty with docker 20.10.x
+	github.com/docker/distribution => github.com/docker/distribution v2.8.2+incompatible // rancher-machine requires a replace is set
+	github.com/docker/docker => github.com/docker/docker v20.10.25+incompatible // rancher-machine requires a repalce is set
 
 	github.com/knative/pkg => github.com/rancher/pkg v0.0.0-20190514055449-b30ab9de040e
 	github.com/matryer/moq => github.com/rancher/moq v0.0.0-20200712062324-13d1f37d2d77
-	github.com/opencontainers/runc => github.com/opencontainers/runc v1.1.5
+	github.com/opencontainers/image-spec => github.com/opencontainers/image-spec v1.1.0-rc2 // needed for containers/image/v5
+	github.com/opencontainers/runc => github.com/opencontainers/runc v1.1.12
 
 	github.com/rancher/machine => github.com/rancher/machine v0.15.0-rancher109
 	github.com/rancher/rancher/pkg/apis => ./pkg/apis
 	github.com/rancher/rancher/pkg/client => ./pkg/client
 
-	go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc => go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc v0.20.0
-	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp => go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.20.0
-	go.opentelemetry.io/otel => go.opentelemetry.io/otel v0.20.0
+	go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc => go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc v0.35.0
+	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp => go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.35.1
 	go.opentelemetry.io/otel/exporters/otlp => go.opentelemetry.io/otel/exporters/otlp v0.20.0
-	go.opentelemetry.io/otel/sdk => go.opentelemetry.io/otel/sdk v0.20.0
-	go.opentelemetry.io/otel/trace => go.opentelemetry.io/otel/trace v0.20.0
-	go.opentelemetry.io/proto/otlp => go.opentelemetry.io/proto/otlp v0.7.0
+	go.opentelemetry.io/otel/metric => go.opentelemetry.io/otel/metric v0.31.0
+	go.opentelemetry.io/otel/sdk => go.opentelemetry.io/otel/sdk v1.10.0
+	go.opentelemetry.io/otel/trace => go.opentelemetry.io/otel/trace v1.10.0
+	go.opentelemetry.io/proto/otlp => go.opentelemetry.io/proto/otlp v0.19.0
 	go.qase.io/client => github.com/rancher/qase-go/client v0.0.0-20231114201952-65195ec001fa
 
-	helm.sh/helm/v3 => github.com/rancher/helm/v3 v3.9.0-rancher1
-	k8s.io/api => k8s.io/api v0.25.12
-	k8s.io/apiextensions-apiserver => k8s.io/apiextensions-apiserver v0.25.12
-	k8s.io/apimachinery => k8s.io/apimachinery v0.25.12
-	k8s.io/apiserver => k8s.io/apiserver v0.25.12
-	k8s.io/cli-runtime => k8s.io/cli-runtime v0.25.12
-	k8s.io/client-go => github.com/rancher/client-go v1.25.4-rancher1
-	k8s.io/cloud-provider => k8s.io/cloud-provider v0.25.12
-	k8s.io/cluster-bootstrap => k8s.io/cluster-bootstrap v0.25.12
-	k8s.io/code-generator => k8s.io/code-generator v0.25.12
-	k8s.io/component-base => k8s.io/component-base v0.25.12
-	k8s.io/component-helpers => k8s.io/component-helpers v0.25.12
-	k8s.io/controller-manager => k8s.io/controller-manager v0.25.12
-	k8s.io/cri-api => k8s.io/cri-api v0.25.12
-	k8s.io/csi-translation-lib => k8s.io/csi-translation-lib v0.25.12
-	k8s.io/kube-aggregator => k8s.io/kube-aggregator v0.25.12
-	k8s.io/kube-controller-manager => k8s.io/kube-controller-manager v0.25.12
-	k8s.io/kube-proxy => k8s.io/kube-proxy v0.25.12
-	k8s.io/kube-scheduler => k8s.io/kube-scheduler v0.25.12
-	k8s.io/kubectl => k8s.io/kubectl v0.25.12
-	k8s.io/kubelet => k8s.io/kubelet v0.25.12
-	k8s.io/kubernetes => k8s.io/kubernetes v1.25.12
-	k8s.io/legacy-cloud-providers => k8s.io/legacy-cloud-providers v0.25.12
-	k8s.io/metrics => k8s.io/metrics v0.25.12
-	k8s.io/mount-utils => k8s.io/mount-utils v0.25.12
-	k8s.io/pod-security-admission => k8s.io/pod-security-admission v0.25.12
-	k8s.io/sample-apiserver => k8s.io/sample-apiserver v0.25.12
+	helm.sh/helm/v3 => github.com/rancher/helm/v3 v3.12.3-rancher1
+	k8s.io/api => k8s.io/api v0.27.4
+	k8s.io/apiextensions-apiserver => k8s.io/apiextensions-apiserver v0.27.4
+	k8s.io/apimachinery => k8s.io/apimachinery v0.27.4
+	k8s.io/apiserver => k8s.io/apiserver v0.27.4
+	k8s.io/cli-runtime => k8s.io/cli-runtime v0.27.4
+	k8s.io/client-go => github.com/rancher/client-go v1.27.4-rancher1
+	k8s.io/cloud-provider => k8s.io/cloud-provider v0.27.4
+	k8s.io/cluster-bootstrap => k8s.io/cluster-bootstrap v0.27.4
+	k8s.io/code-generator => k8s.io/code-generator v0.27.4
+	k8s.io/component-base => k8s.io/component-base v0.27.4
+	k8s.io/component-helpers => k8s.io/component-helpers v0.27.4
+	k8s.io/controller-manager => k8s.io/controller-manager v0.27.4
+	k8s.io/cri-api => k8s.io/cri-api v0.27.4
+	k8s.io/csi-translation-lib => k8s.io/csi-translation-lib v0.27.4
+	k8s.io/dynamic-resource-allocation => k8s.io/dynamic-resource-allocation v0.27.4
+	k8s.io/kms => k8s.io/kms v0.27.4
+	k8s.io/kube-aggregator => k8s.io/kube-aggregator v0.27.4
+	k8s.io/kube-controller-manager => k8s.io/kube-controller-manager v0.27.4
+	k8s.io/kube-openapi => k8s.io/kube-openapi v0.0.0-20230501164219-8b0f38b5fd1f
+	k8s.io/kube-proxy => k8s.io/kube-proxy v0.27.4
+	k8s.io/kube-scheduler => k8s.io/kube-scheduler v0.27.4
+	k8s.io/kubectl => k8s.io/kubectl v0.27.4
+	k8s.io/kubelet => k8s.io/kubelet v0.27.4
+	k8s.io/kubernetes => k8s.io/kubernetes v1.27.4
+	k8s.io/legacy-cloud-providers => k8s.io/legacy-cloud-providers v0.27.4
+	k8s.io/metrics => k8s.io/metrics v0.27.4
+	k8s.io/mount-utils => k8s.io/mount-utils v0.27.4
+	k8s.io/pod-security-admission => k8s.io/pod-security-admission v0.27.4
+	k8s.io/sample-apiserver => k8s.io/sample-apiserver v0.27.4
+
+	oras.land/oras-go => oras.land/oras-go v1.2.2 // for docker 20.10.x compatibility
 
 	sigs.k8s.io/aws-iam-authenticator => github.com/rancher/aws-iam-authenticator v0.5.9-0.20220713170329-78acb8c83863
-	sigs.k8s.io/cluster-api => sigs.k8s.io/cluster-api v1.2.12
+	sigs.k8s.io/cluster-api => sigs.k8s.io/cluster-api v1.5.0
 )
 
 require (
