@@ -1,34 +1,35 @@
 module github.com/rancher/rancher/pkg/apis
 
-go 1.19
+go 1.20
 
 // wrangler bracnhes need to be updated before replace can be removed
-replace github.com/rancher/wrangler v1.1.1 => github.com/rancher/wrangler v1.1.1-0.20230629203936-0e36fee7aea5
+replace github.com/rancher/wrangler v1.1.1 => github.com/rancher/wrangler v1.1.1-0.20230831050635-df1bd5aae9df
 
-replace k8s.io/client-go => github.com/rancher/client-go v1.25.4-rancher1
+replace k8s.io/client-go => github.com/rancher/client-go v1.27.4-rancher1
 
 require (
 	github.com/rancher/aks-operator v1.1.3
 	github.com/rancher/eks-operator v1.2.2
-	github.com/rancher/fleet/pkg/apis v0.0.0-20230821130910-23e1c146755a
+	github.com/rancher/fleet/pkg/apis v0.0.0-20230901075223-437edb7091f5
 	github.com/rancher/gke-operator v1.1.6
-	github.com/rancher/norman v0.0.0-20230811152901-078862e5648c
-	github.com/rancher/rke v1.4.10
+	github.com/rancher/norman v0.0.0-20240207153035-cb54924f25c7
+	github.com/rancher/rke v1.4.15
 	github.com/rancher/wrangler v1.1.1
 	github.com/sirupsen/logrus v1.9.3
-	k8s.io/api v0.25.12
-	k8s.io/apimachinery v0.25.12
-	sigs.k8s.io/cluster-api v1.2.12
+	k8s.io/api v0.27.4
+	k8s.io/apimachinery v0.27.4
+	sigs.k8s.io/cluster-api v1.5.0
 )
 
 require (
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/blang/semver v3.5.1+incompatible // indirect
+	github.com/blang/semver/v4 v4.0.0 // indirect
 	github.com/cespare/xxhash/v2 v2.2.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/emicklei/go-restful/v3 v3.10.1 // indirect
-	github.com/evanphx/json-patch v4.12.0+incompatible // indirect
-	github.com/fsnotify/fsnotify v1.5.4 // indirect
+	github.com/evanphx/json-patch/v5 v5.6.0 // indirect
+	github.com/fsnotify/fsnotify v1.6.0 // indirect
 	github.com/ghodss/yaml v1.0.0 // indirect
 	github.com/go-logr/logr v1.2.4 // indirect
 	github.com/go-openapi/jsonpointer v0.19.6 // indirect
