@@ -194,7 +194,7 @@ func (r *RKE2ConfigServer) connectConfigYaml(name, ns string, rw http.ResponseWr
 		return
 	}
 
-	kubernetesVersion, err := r.getClusterKubernetesVersion(mpSecret.Labels[capi.ClusterLabelName], ns)
+	kubernetesVersion, err := r.getClusterKubernetesVersion(mpSecret.Labels[capi.ClusterNameLabel], ns)
 	if err != nil {
 		http.Error(rw, err.Error(), http.StatusInternalServerError)
 		return
@@ -273,7 +273,7 @@ func (r *RKE2ConfigServer) infoKubernetesVersion(machineID, ns string) (string, 
 		return "", err
 	}
 
-	clusterName, ok := machine.Labels[capi.ClusterLabelName]
+	clusterName, ok := machine.Labels[capi.ClusterNameLabel]
 	if !ok {
 		return "", fmt.Errorf("unable to find cluster name for machine")
 	}
@@ -414,9 +414,11 @@ func (r *RKE2ConfigServer) findSA(req *http.Request) (string, *corev1.Secret, er
 	}
 
 	logrus.Debugf("[rke2configserver] %s/%s starting token secret watch for planSA %s/%s", machineNamespace, machineName, planSA.Namespace, planSA.Name)
-	// start watch for the planSA corresponding secret, using a field selector.
+	// start watch for the planSA corresponding secret, using a label selector.
 	respSecret, err := r.secrets.Watch(machineNamespace, metav1.ListOptions{
-		FieldSelector: fmt.Sprintf("metadata.name=%s", serviceaccounttoken.ServiceAccountSecretName(planSA)),
+		LabelSelector: labels.Set{
+			serviceaccounttoken.ServiceAccountSecretLabel: planSA.Name,
+		}.String(),
 	})
 	if err != nil {
 		return "", nil, err
