@@ -15,7 +15,7 @@ const (
 type PersistentVolumeClaimSpec struct {
 	AccessModes      []string                   `json:"accessModes,omitempty" yaml:"accessModes,omitempty"`
 	DataSource       *TypedLocalObjectReference `json:"dataSource,omitempty" yaml:"dataSource,omitempty"`
-	DataSourceRef    *TypedLocalObjectReference `json:"dataSourceRef,omitempty" yaml:"dataSourceRef,omitempty"`
+	DataSourceRef    *TypedObjectReference      `json:"dataSourceRef,omitempty" yaml:"dataSourceRef,omitempty"`
 	Resources        *ResourceRequirements      `json:"resources,omitempty" yaml:"resources,omitempty"`
 	Selector         *LabelSelector             `json:"selector,omitempty" yaml:"selector,omitempty"`
 	StorageClassName string                     `json:"storageClassName,omitempty" yaml:"storageClassName,omitempty"`
