@@ -37,7 +37,7 @@ type PersistentVolumeClaim struct {
 	Created              string                       `json:"created,omitempty" yaml:"created,omitempty"`
 	CreatorID            string                       `json:"creatorId,omitempty" yaml:"creatorId,omitempty"`
 	DataSource           *TypedLocalObjectReference   `json:"dataSource,omitempty" yaml:"dataSource,omitempty"`
-	DataSourceRef        *TypedLocalObjectReference   `json:"dataSourceRef,omitempty" yaml:"dataSourceRef,omitempty"`
+	DataSourceRef        *TypedObjectReference        `json:"dataSourceRef,omitempty" yaml:"dataSourceRef,omitempty"`
 	Labels               map[string]string            `json:"labels,omitempty" yaml:"labels,omitempty"`
 	Name                 string                       `json:"name,omitempty" yaml:"name,omitempty"`
 	NamespaceId          string                       `json:"namespaceId,omitempty" yaml:"namespaceId,omitempty"`
