@@ -334,7 +334,7 @@ func (n *nsLifecycle) reconcileNamespaceProjectClusterRole(ns *v1.Namespace) err
 			return err
 		}
 
-		roleCli := n.m.workload.RBAC.ClusterRoles("")
+		roleCli := n.m.clusterRoles
 		nsInDesiredRole := false
 		for _, c := range clusterRoles {
 			cr, ok := c.(*rbacv1.ClusterRole)
@@ -442,7 +442,7 @@ func (n *nsLifecycle) reconcileNamespaceProjectClusterRole(ns *v1.Namespace) err
 }
 
 func (m *manager) createProjectNSRole(roleName, verb, ns, projectName string) error {
-	roleCli := m.workload.RBAC.ClusterRoles("")
+	roleCli := m.clusterRoles
 
 	cr := &rbacv1.ClusterRole{
 		ObjectMeta: metav1.ObjectMeta{
