@@ -384,9 +384,9 @@ func machineDeployments(cluster *rancherv1.Cluster, capiCluster *capi.Cluster, d
 				Template: capi.MachineTemplateSpec{
 					ObjectMeta: capi.ObjectMeta{
 						Labels: map[string]string{
-							capi.ClusterLabelName:           capiCluster.Name,
+							capi.ClusterNameLabel:           capiCluster.Name,
 							capr.ClusterNameLabel:           capiCluster.Name,
-							capi.MachineDeploymentLabelName: machineDeploymentName,
+							capi.MachineDeploymentNameLabel: machineDeploymentName,
 							capr.RKEMachinePoolNameLabel:    machinePool.Name,
 						},
 						Annotations: machineSpecAnnotations,
@@ -419,7 +419,7 @@ func machineDeployments(cluster *rancherv1.Cluster, capiCluster *capi.Cluster, d
 
 		if machinePool.ControlPlaneRole {
 			machineDeployment.Spec.Template.Labels[capr.ControlPlaneRoleLabel] = "true"
-			machineDeployment.Spec.Template.Labels[capi.MachineControlPlaneLabelName] = "true"
+			machineDeployment.Spec.Template.Labels[capi.MachineControlPlaneLabel] = "true"
 		}
 
 		if machinePool.WorkerRole {
@@ -476,7 +476,7 @@ func deploymentHealthChecks(machineDeployment *capi.MachineDeployment, machinePo
 			ClusterName: machineDeployment.Spec.ClusterName,
 			Selector: metav1.LabelSelector{ // this health check only applies to machines in this deployment
 				MatchLabels: map[string]string{
-					capi.MachineDeploymentLabelName: machineDeployment.Name,
+					capi.MachineDeploymentNameLabel: machineDeployment.Name,
 				},
 			},
 			UnhealthyConditions: []capi.UnhealthyCondition{ // if a node status is unready or unknown for the timeout mark it unhealthy
